organize model directory
