+++ app/src/androidTest/java/com/john/shopper/SettingsTest.java
@@ -1,7 +1,5 @@
 package com.john.shopper;
 
-import static androidx.test.espresso.Espresso.openActionBarOverflowOrOptionsMenu;
-
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
@@ -10,12 +8,9 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
-import androidx.test.espresso.matcher.PreferenceMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
-
-import com.john.shopper.model.SettingsModel;
 
 import org.junit.After;
 import org.junit.Before;
