+++ app/src/androidTest/java/com/john/shopper/ShoppingListItemsTest.java
@@ -19,9 +19,9 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
-import com.john.shopper.model.jsonModel.ShoppingList;
-import com.john.shopper.model.jsonModel.ShoppingListItem;
-import com.john.shopper.model.jsonModel.JSONModel;
+import com.john.shopper.model.ShoppingList;
+import com.john.shopper.model.ShoppingListItem;
+import com.john.shopper.model.JSONModel;
 
 import org.junit.After;
 import org.junit.Before;
