+++ app/src/main/java/com/john/shopper/model/ShoppingListItem.java
@@ -1,36 +1,20 @@
 package com.john.shopper.model;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.room.ColumnInfo;
-import androidx.room.Entity;
-import androidx.room.PrimaryKey;
+import java.io.Serializable;
+import java.util.UUID;
 
-@Entity(tableName = "items")
-public class ShoppingListItem {
-
-    @PrimaryKey(autoGenerate = true)
-    public long id;
-
-    @ColumnInfo(name = "list_id")
-    public long listId;
-
-    @ColumnInfo(name = "name")
+public class ShoppingListItem implements Serializable {
+    public String itemId;
     public String name;
-
     public int quantity;
-
-    @ColumnInfo(name = "is_complete")
     public boolean isComplete;
-
-    @ColumnInfo(name = "is_section")
     public boolean isSection;
 
-    public int position;
-
-    @NonNull
-    @Override
-    public String toString() {
-        return this.name;
+    public ShoppingListItem(String name, int quantity, boolean isComplete, boolean isSection) {
+        this.itemId = UUID.randomUUID().toString();
+        this.name = name;
+        this.quantity = quantity;
+        this.isComplete = isComplete;
+        this.isSection = isSection;
     }
 }
