+++ app/src/main/java/com/john/shopper/model/old/ShoppingListDAO.java
@@ -1,10 +1,9 @@
-package com.john.shopper.model;
+package com.john.shopper.model.old;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.Query;
-import androidx.room.Update;
 
 import java.util.List;
